requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
supabase==2.9.1
python-dotenv==1.0.1
lxml==5.3.0
